requirements.txt
pandas
scikit-learn
pyyaml
mlflow
fastapi
uvicorn
pydantic
gradio
evidently
great_expectations
pytest